Fixed stupid bugs

+++ Windows OEM Editor/MainWindow.xaml.cs
@@ -105,6 +105,8 @@ namespace Windows_OEM_Editor
                             Path.Combine(Path.GetPathRoot(Environment.GetFolderPath(Environment.SpecialFolder.System)),
                                 "oemImage.bmp");
 
+                        if (tb.ToolTip.ToString() == f) return;
+
                         if(File.Exists(f)) File.Delete(f);
 
                         image_.Save(f, ImageFormat.Bmp);
